PeriDyn.BondBasedSpecific — Type

BondBasedSpecific

Bond based specific material type.

Fields

bond_stiffness::AbstractArray{T,2}: Bond stiffness matrix.

bulk_modulus::AbstractArray{T,2}: Bulk modulus matrix.

critical_stretch::AbstractArray{T,2}: Critical stretch matrix.

density::AbstractArray{T,1}: Density vector.

func::Function: Bond force function.

Constructor

BondBasedSpecific(K::AbstractArray, critical_stretch::AbstractArray, density::AbstractArray; horizon=nothing, func=nothing)
BondBasedSpecific(K::Real, critical_stretch::Real, density::Real; kwargs...)

PeriDyn.BondBasedSpecific — Method

BondBasedSpecific(K::AbstractArray, critical_stretch::AbstractArray, density::AbstractArray; horizon=nothing, func=nothing)

Constructs BondBasedSpecific material type.

Arguments

K::AbstractArray: Bulk modulus matrix.

critical_stretch::AbstractArray: Critical stretch matrix.

density::AbstractArray: Density vector.

horizon::Real: Horizon.

func::Function: Bond force function.

Returns

BondBasedSpecific: Bond based specific material type.

PeriDyn.BoundaryCondition — Type

Abstract class for boundary conditions.

PeriDyn.BoundaryConditionat0 — Type

Abstract class for boundary conditions that are applied at time 0.

PeriDyn.DeltaScaleBC — Type

Struct representing the DeltaScaleBC boundary condition.

PeriDyn.DeltaScaleBC — Method

DeltaScaleBC(bool, scale, fixpoint; onlyatstart=false)

Construct a DeltaScaleBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

scale: Scale factor applied to the elements.

fixpoint: Reference point used for scaling.

onlyatstart: Flag indicating if the boundary condition is applied only at the start (default: false).

Returns

A DeltaScaleBC object representing the boundary condition.

PeriDyn.ElastoPlasticSolidMaterial — Type

ElastoPlasticSolidMaterial(type::UnitRange{Int64}, general::GeneralMaterial, specific::ElastoPlasticSolidSpecific)

PeriDyn.ElastoPlasticSolidSpecific — Type

Specific Elasto Plastic Solid Material type.

PeriDyn.ElastoPlasticSolidSpecific — Method

ElastoPlasticSolidSpecific(bulk_modulus::Array{Float64, 1}, shear_modulus::Array{Float64,1}, critical_stretch::Array{Float64,1}, density::Array{Float64,1})

Specific Elasto Plastic Solid Material type.

PeriDyn.ElastoPlasticSolidSpecificFull — Type

Specific Elasto Plastic Solid Material type.

PeriDyn.FixBC — Type

Struct representing the FixBC boundary condition.

PeriDyn.FixBC — Method

FixBC(bool; onlyatstart=false)

Construct a FixBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

onlyatstart: Flag indicating if the boundary condition is applied only at the start (default: false).

Returns

A FixBC object representing the boundary condition.

PeriDyn.GeneralMaterial — Type

GeneralMaterial

General peridynamics material type.

Fields

y::AbstractArray{Float64,2}: Displacement of the material.

velocity::AbstractArray{Float64,2}: Velocity of the material.

x::AbstractArray{Float64,2}: Position of the material.

volume::AbstractArray{Float64,1}: Volume of the material.

type::AbstractArray{Int64,1}: Type of the material.

particle_size::Float64: Particle size of the material.

horizon::Float64: Horizon of the material.

family::AbstractArray{Int64,2}: Family of the material.

intact::AbstractArray{Bool, 2}: Intact of the material.

weighted_volume::AbstractArray{Float64,1}: Weighted volume of the material.

deformed::AbstractVector{Bool}: Deformed of the material.

skip_bb::Bool: Skip bond based material.

skip_bb ? why do we need this?

PeriDyn.GeneralMaterial — Method

GeneralMaterial(y0, v0, x, volume, type, horizon; max_neigh=100, particle_size=0, skip_bb=false)

Creates a `GeneralMaterial` type.

Arguments

y0::AbstractArray{Float64,2}: Initial displacement of the material.

v0::AbstractArray{Float64,2}: Initial velocity of the material.

x::AbstractArray{Float64,2}: Initial position of the material.

volume::AbstractArray{Float64,1}: Volume of the material.

type::AbstractArray{Int64,1}: Type of the material.

horizon::Float64: Horizon of the material.

Keyword Arguments

max_neigh::Int64 = 100: Maximum number of neighbors.

particle_size::Float64 = 0: Particle size of the material.

skip_bb::Bool = false: Skip bond based material.

Returns

mat::GeneralMaterial: General material type.

PeriDyn.GeneralMaterial — Method

GeneralMaterial(items::Dict, args...; kwargs...)

Creates a `GeneralMaterial` type. `items` is a dictionary of the fields of the `GeneralMaterial` type
typical comes from PDMaterialPoints package.

Arguments

items::Dict: Dictionary of the fields of the GeneralMaterial type.

args...: Arguments of the GeneralMaterial type.

kwargs...: Keyword arguments of the GeneralMaterial type.

Returns

mat::GeneralMaterial: General material type.

see also

GeneralMaterial(y0, v0, x, volume, type, horizon; max_neigh=100, particle_size=0, skip_bb=false)

PeriDyn.NonLinearRepulsionModel11 — Type

NonLinearRepulsionModel11(type::UnitRange{Int64}, bid::Int64, material::GeneralMaterial, exponent::Float64, stifness::Float64, neighs::AbstractArray{Int64, 2}, distance::Float64, max_neighs::Int64)

Nonlinear repulsive model for 1-1 material blocks.

PeriDyn.NonLinearRepulsionModel12 — Type

NonLinearRepulsionModel12(pair::Vector{UnitRange{Int64}}, exponent::Float64, stifness::Float64, equi_dist::Float64, neighs::AbstractArray{Int64, 2}, distance::Float64, max_neighs::Int64)

Nonlinear repulsive model for 1-2 material blocks.

PeriDyn.OrdinaryStateBasedMaterial — Type

OrdinaryStateBasedMaterial(type::UnitRange{Int64}, general::GeneralMaterial, specific::OrdinaryStateBasedSpecific)

PeriDyn.OrdinaryStateBasedSpecific — Method

OrdinaryStateBasedSpecific(bulk_modulus::AbstractArray{Float64, 1}, shear_modulus::AbstractArray{Float64,1}, critical_stretch::AbstractArray{Float64,1}, density::AbstractArray{Float64,1})

Specific ordinary state based materail type.

PeriDyn.PairwiseNNMaterial — Type

Bond based material type (NN).

PeriDyn.PairwiseNNSpecific — Type

Specific bond based material type (NN) .

PeriDyn.PeridynamicsMaterial — Type

PeridynamicsMaterial

Abstract Peridynamics material type.

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(name, type, bid, gen, spc)

Creates a peridynamics material type with given name, type, block id, general material type
and specific material type.

Arguments

name::String: Name of the material.

type::Union{UnitRange{Int64}, AbstractVector{Int64}}: Type of the material.

bid::Int64: Block id of the material.

gen::GeneralMaterial: General material type.

spc::SpecificMaterial: Specific material type.

Returns

mat::PeridynamicsMaterial: Peridynamics material type.

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(gen, spc::ElastoPlasticSolidSpecific)

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(gen, spc::SkipSpecific)

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(bid, gen, spc; name="PeriMat")

Creates a peridynamics material type with given block id, general and specific material type. `name` is optional.

Arguments

bid::Int64: Block id of the material.

gen::GeneralMaterial: General material type.

spc::SpecificMaterial: Specific material type.

name::String: Name of the material.

Returns

mat::PeridynamicsMaterial: Peridynamics material type.

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(gen, spc; name="PeriMat")

Creates a peridynamics material type with given general and specific material type. `name` is optional.

Arguments

gen::GeneralMaterial: General material type.

spc::SpecificMaterial: Specific material type.

name::String: Name of the material.

Returns

mat::PeridynamicsMaterial: Peridynamics material type.

PeriDyn.RepulsionModel11 — Type

RepulsionModel11

Abstract type for repulsion model for a single material type.

PeriDyn.RepulsionModel12 — Type

RepulsionModel12

Abstract type for repulsion model for two material types.

PeriDyn.ScaleFixWaitBC — Type

Structure representing a ScaleFixWaitBC boundary condition.

PeriDyn.ScaleFixWaitBC — Method

ScaleFixWaitBC(bool, scale, fixpoint, wait, scalebool; applyafter=0, onlyatstart=false)

Construct a ScaleFixWaitBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

scale: Scale factor for the elements.

fixpoint: Fix point for the elements.

wait: Number of time steps to wait before applying the condition.

scalebool: Boolean array specifying the elements to be scaled.

applyafter: Number of time steps after which the condition is applied (default: 0).

onlyatstart: Boolean indicating whether the condition is applied only at the start (default: false).

Returns

An instance of ScaleFixWaitBC boundary condition.

PeriDyn.SkipMaterial — Type

SkipMaterial(type::UnitRange{Int64}, general::GeneralMaterial, specific::SkipSpecific)

PeriDyn.SkipSpecific — Type

Specific skip materail type.

PeriDyn.SkipSpecific — Method

Specific skip materail type.

PeriDyn.SpecificMaterial — Type

SpecificMaterial

Abstract specific material type.

PeriDyn.ToFroBC — Type

Struct representing the ToFroBC boundary condition.

PeriDyn.ToFroBC — Method

ToFroBC(bool, rate, freq; applyafter=0, onlyatstart=false)

Construct a ToFroBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

rate: Rate at which the elements move.

freq: Frequency at which the direction of movement changes.

applyafter: Number of steps after which the frequency is applied (default: 0).

onlyatstart: Flag indicating if the boundary condition is applied only at the start (default: false).

Returns

A ToFroBC object representing the boundary condition.

PeriDyn.Env — Method

Env(id::Int64,materials,short_range_repulsion,boundary_conds,dt;state=2)

Create a GeneralEnv for holding parameters for a simulation.

PeriDyn.LJRepulsionModel — Method

LJRepulsionModel(alpha::Float64,epsilon::Float64, mat1::PeridynamicsMaterial,mat2::PeridynamicsMaterial;distanceX=5,max_neighs=50)

LJ repulsive model for 1-2 material blocks.

PeriDyn.LJRepulsionModel — Method

LJRepulsionModel(alpha::Float64,epsilon::Float64, mat1::PeridynamicsMaterial;distanceX=5,max_neighs=50)

LJ repulsive model for 1-1 material blocks.

PeriDyn.LinearRepulsionModel — Method

LinearRepulsionModel(stiffness::Float64, mat1::PeridynamicsMaterial, mat2::PeridynamicsMaterial; distanceX=5, max_neighs=50)

Constructs a linear repulsive model for 1-2 material blocks. The constructors internally call the NonLinearRepulsionModel constructor with a nonlinearity parameter of 1 and the provided arguments.

Arguments:

stiffness: The stiffness coefficient for the repulsive model.

mat1: PeridynamicsMaterial representing the first material block.

mat2: PeridynamicsMaterial representing the second material block.

distanceX: Optional keyword argument specifying the distance parameter (default: 5).

max_neighs: Optional keyword argument specifying the maximum number of neighbors (default: 50).

Returns:

An instance of the LinearRepulsionModel.

PeriDyn.LinearRepulsionModel — Method

LinearRepulsionModel(stiffness::Float64, mat1::PeridynamicsMaterial; distanceX=5, max_neighs=50)

Constructs a linear repulsive model for 1-1 material blocks. The constructors internally call the NonLinearRepulsionModel constructor with a nonlinearity parameter of 1 and the provided arguments.

Arguments:

stiffness: The stiffness coefficient for the repulsive model.

mat1: PeridynamicsMaterial representing the material block.

distanceX: Optional keyword argument specifying the distance parameter (default: 5).

max_neighs: Optional keyword argument specifying the maximum number of neighbors (default: 50).

Returns:

An instance of the LinearRepulsionModel.

PeriDyn.MoveBC — Method

MoveBC(bool, rate; kwargs...)

Create a MoveBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

rate: Rate at which the elements move.

kwargs: Additional keyword arguments passed to ToFroBC.

Returns

A MoveBC object representing the boundary condition. All MoveBC objects are ToFroBC objects with frequency Inf.

PeriDyn.NonLinearRepulsionModel — Method

NonLinearRepulsionModel(exponent::Float64, stifness::Float64, mat1::PeridynamicsMaterial, mat2::PeridynamicsMaterial; distanceD=1.0, distanceX=3.0, max_neighs=50)

Constructs a nonlinear repulsive model for 1-2 material blocks.

Arguments:

exponent: The exponent for the repulsive force calculation.

stifness: The stiffness coefficient for the repulsive force calculation.

mat1: The first PeridynamicsMaterial.

mat2: The second PeridynamicsMaterial.

distanceD: The distance factor for determining the search distance of neighbors within the same material.

distanceX: The distance factor for determining the search distance of neighbors between different materials.

max_neighs: The maximum number of neighbors to consider.

Returns: A NonLinearRepulsionModel12 object.

PeriDyn.NonLinearRepulsionModel — Method

NonLinearRepulsionModel(exponent::Float64, stifness::Float64, mat1::PeridynamicsMaterial; distanceX=5, max_neighs=50)

Constructs a nonlinear repulsive model for 1-1 material blocks.

Arguments:

exponent: The exponent for the repulsive force calculation.

stifness: The stiffness coefficient for the repulsive force calculation.

mat1: The PeridynamicsMaterial.

distanceX: The distance factor for determining the search distance of neighbors.

max_neighs: The maximum number of neighbors to consider.

Returns: A NonLinearRepulsionModel11 object.

PeriDyn.RepulsionModel11_gcal — Method

RepulsionModel11_gcal(mat1, distanceX, max_neighs)

Calculate the parameters for RepulsionModel11 based on the input material, distance, and maximum neighbors.

Arguments:

mat1: The material (type RepulsionModel11).

distanceX: The distance factor.

max_neighs: The maximum number of neighbors.

Returns a tuple containing the calculated parameters for RepulsionModel11.

PeriDyn.RepulsionModel12_gcal — Method

RepulsionModel12_gcal(mat1, mat2, distanceX, max_neighs)

Calculate the parameters for RepulsionModel12 based on the input materials, distance, and maximum neighbors.

Arguments:

mat1: The first material (type RepulsionModel11).

mat2: The second material (type RepulsionModel11).

distanceX: The distance factor.

max_neighs: The maximum number of neighbors.

Returns a tuple containing the calculated parameters for RepulsionModel12.

PeriDyn.ShortRangeRepulsionModel — Method

ShortRangeRepulsionModel(spring_const::Float64, mat1::PeridynamicsMaterial, mat2::PeridynamicsMaterial; distanceX=5, max_neighs=50)

Create a ShortRangeRepulsionModel for 1-2 material blocks.

Arguments:

spring_const: Float64, the spring constant for repulsion.

mat1: PeridynamicsMaterial, the first material block.

mat2: PeridynamicsMaterial, the second material block.

distanceX: Float64, the distance factor for neighbor searching. Default is 5.

max_neighs: Int64, the maximum number of neighbors. Default is 50.

Returns:

ShortRangeRepulsionModel12: The ShortRangeRepulsionModel for 1-2 material blocks.

PeriDyn.ShortRangeRepulsionModel — Method

ShortRangeRepulsionModel(spring_const::Float64, mat1::PeridynamicsMaterial; distanceX=5, max_neighs=50)

Create a ShortRangeRepulsionModel for 1-1 material blocks.

Arguments:

spring_const: Float64, the spring constant for repulsion.

mat1: PeridynamicsMaterial, the material block.

distanceX: Float64, the distance factor for neighbor searching. Default is 5.

max_neighs: Int64, the maximum number of neighbors. Default is 50.

Returns:

ShortRangeRepulsionModel11: The ShortRangeRepulsionModel for 1-1 material blocks.

PeriDyn._cudaconvert — Method

_cudaconvert(x::T) where T <: Union{RepulsionModel11,RepulsionModel12}

Converts a single RepulsionModel11 or RepulsionModel12 object to a CUDA-compatible type.

PeriDyn._cudaconvert — Method

_cudaconvert(x::Vector{T}) where T <: Union{RepulsionModel11,RepulsionModel12}

Converts a vector of RepulsionModel11 or RepulsionModel12 objects to CUDA-compatible types.

PeriDyn._tij — Method

_tij(x,mi,thetai,eij,K,G)::Float64

calculates tij force density magnitude (actually acceleration).

PeriDyn.apply_bc! — Method

apply_bc!(env, BC::BoundaryCondition, on::Symbol)

Apply the specified boundary condition BC to the given environment env on the specified aspect on.

Arguments

env: The environment to which the boundary condition is applied.

BC: The boundary condition to apply.

on::Symbol: The aspect on which the boundary condition is applied (:position or :velocity).

PeriDyn.apply_bc! — Method

apply_bc!(env, BC::BoundaryCondition, ::Type{Val{:position}})

Apply the general boundary condition BC to the position aspect of the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The general boundary condition to apply.

PeriDyn.apply_bc! — Method

apply_bc!(env, BC::BoundaryCondition, ::Type{Val{:velocity}})

Apply the general boundary condition BC to the velocity aspect of the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The general boundary condition to apply.

PeriDyn.apply_bc_at0! — Method

apply_bc_at0!(env, BC::DeltaScaleBC)

Apply the DeltaScaleBC boundary condition at time 0 to the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The DeltaScaleBC boundary condition to apply.

PeriDyn.apply_bc_at0! — Method

apply_bc_at0!(env, BC::FixBC)

Apply the FixBC boundary condition at time 0 to the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The FixBC boundary condition to apply.

PeriDyn.apply_bc_at0! — Method

apply_bc_at0!(env, BC::ScaleFixWaitBC)

Apply the ScaleFixWaitBC boundary condition at time 0.

Arguments

env: Environment in which the condition is applied.

BC: Instance of ScaleFixWaitBC boundary condition.

PeriDyn.apply_bc_at0! — Method

apply_bc_at0!(env, BC::ToFroBC)

Apply the ToFroBC boundary condition at time 0 to the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The ToFroBC boundary condition to apply.

PeriDyn.cal_family! — Method

cal_family!(family::Array{Int64,2},x::Array{Float64,2}, horizon::Float64)

Calculate family members for each material point (inplace).

PeriDyn.cal_family — Method

cal_family!(family::Array{Int64,2},x::Array{Float64,2}, horizon::Float64)

Calculate family members for each material point.

PeriDyn.cell_number — Method

cell_number(i,j,k,N)

Calculates cell number for given i,j,k cell indices (when in a list).

PeriDyn.check! — Method

check!(BC::DeltaScaleBC, env)

Perform a check on the DeltaScaleBC boundary condition.

Arguments

BC: The DeltaScaleBC boundary condition to check.

env: The environment associated with the boundary condition.

PeriDyn.check! — Method

check!(BC::ScaleFixWaitBC, env)

Check if the ScaleFixWaitBC boundary condition needs to be applied.

Arguments

BC: Instance of ScaleFixWaitBC boundary condition.

env: Environment in which the condition is applied.

PeriDyn.check! — Method

check!(BC::ToFroBC, env)

Perform a check on the ToFroBC boundary condition.

Arguments

BC: The ToFroBC boundary condition to check.

env: The environment associated with the boundary condition.

PeriDyn.collision_box — Method

collision_box(x1::Array{Float64,2}, x2::Array{Float64,2}, skin::Float64)

Calculates collision box between two material blocks.

Arguments

x1: Positions of material points (block 1)

x2: Positions of material points (block 2)

skin: Extra distance to consider (usually >= particle size)

Output

box_min: Minimum position limits for overlap

box_max: Maximum position limits for overlap

ifoverlap: Boolean indicating if there is an overlap

PeriDyn.dilatation! — Method

dilatation(y, x, intact, family, volume, m, particle_size, horizon, ::Type{Val{:cuda}})

It gives dilatation as given ordinary state material model.

PeriDyn.dilatation! — Method

dilatation(y, x, intact, family, volume, m, particle_size, horizon, device::Type{Val{:cpu}})

It gives dilatation as given ordinary state material model.

PeriDyn.dilatation — Method

dilatation(y, x, intact, family, volume, m, particle_size, horizon, device::Type{Val{:cpu}})

It gives dilatation as given ordinary state material model.

PeriDyn.dilatation — Method

dilatation(y, x, intact, family, volume, m, particle_size, horizon, device::Type{Val{:cuda}})

It gives dilatation as given ordinary state material model.

PeriDyn.force_density — Method

force_density_T(f, y, limits, mat::PeridynamicsMaterial)

Calculates the force of the material.

Arguments

f::AbstractArray{Float64,2}: Force of the material.

y::AbstractArray{Float64,2}: Displacement of the material.

limits::AbstractVector{Int64}: Limits of the material.

mat::PeridynamicsMaterial: Peridynamics material type.

Returns (may be in-place)

f::AbstractArray{Float64,2}: Force of the material.

PeriDyn.force_density_T — Method

force_density_T(force::AbstractArray{Float64,2}, y::AbstractArray{Float64,2}, limits, mat::BondBasedMaterial, ::Type{Val{:cuda}}; particles=nothing)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(force::AbstractArray{Float64,2}, y::AbstractArray{Float64,2}, limits, mat::OrdinaryStateBasedMaterial, device::Type{Val{:cuda}}; particles=nothing)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(mat::SkipMaterial)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(mat::BondBasedMaterial)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(mat::BondBasedMaterial)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(f, y, limits, mat::PeridynamicsMaterial, device::Symbol)

Calculates the force of the material.

Arguments

f::AbstractArray{Float64,2}: Force of the material.

y::AbstractArray{Float64,2}: Displacement of the material.

limits::AbstractVector{Int64}: Limits of the material.

mat::PeridynamicsMaterial: Peridynamics material type.

device::Symbol: Device of the material.

Returns (may be in-place)

f::AbstractArray{Float64,2}: Force of the material.

PeriDyn.force_density_T — Method

force_density_T(f, y, limits, mat::GeneralMaterial, device::Type{Val{:cpu}}; particles=nothing)

Calculates the force of the material.

Arguments

f::AbstractArray{Float64,2}: Force of the material.

y::AbstractArray{Float64,2}: Displacement of the material.

limits::AbstractVector{Int64}: Limits of the material.

mat::GeneralMaterial: General material type.

device::Type{Val{:cpu}}: Device of the material.

particles::Union{Nothing, AbstractVector{Int64}}: Particles to calculate the force.

Returns (may be in-place)

f::AbstractArray{Float64,2}: Force of the material.

PeriDyn.force_density_T — Method

force_density_T(y::Array{Float64,2},mat::ElastoPlasticSolidMaterial)

Calculates force density (actually acceleration) for ordinary state based material type.

PeriDyn.force_density_T — Method

force_density_T(mat::PairwiseNNMaterial)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(mat::PairwiseNNSpecific)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.get_cells — Method

get_cells(x::Array{Float64,2},horizon::Float64)

Fill cells with material points.

PeriDyn.get_repulsion_force_fn — Method

get_repulsion_force_fn(RepMod::NonLinearRepulsionModel11)

Returns a repulsion force function for a NonLinearRepulsionModel11 object.

Arguments:

RepMod: The NonLinearRepulsionModel11 object.

Returns: A repulsion force function that takes a distance vector and returns the repulsive acceleration.

PeriDyn.horizon_correction — Method

horizon_correction(dr, ps, hr)

It gives horizon correction factor (It will give 1 as of now).

PeriDyn.influence_function — Method

influence_function(dr)

It gives influence function factor (It will give 1/r as of now).

PeriDyn.init — Method

init(spc::T1, gen::T2) where {T1<:SpecificMaterial, T2<:GeneralMaterial}

Initialize specific material type. It uses information from GeneralMaterial to create SpecificMaterial type.

Arguments

spc::T1: Specific material type.

gen::T2: General material type.

Returns

spc::T1: Initialized specific material type.

PeriDyn.jld2array — Method

jld2array(file, N; start=0, step=100)

Loads data from JLD files into an array.

Arguments:

file: String, the base name of the JLD files.

N: Int, the number of files to load.

start: Int, the index of the first file to load. Default is 0.

step: Int, the step size between files to load. Default is 100.

Returns:

Array, an array containing the data loaded from the JLD files.

Note: This function iterates over a range of files and loads data from each JLD file using the jldread function.

PeriDyn.jld2ovito — Method

jld2ovito(file, N; start=0, step=100)

Converts JLD files to Ovito data files.

Arguments:

file: String, the base name of the JLD files.

N: Int, the number of files to convert.

start: Int, the index of the first file to convert. Default is 0.

step: Int, the step size between files to convert. Default is 100.

Note: This function iterates over a range of files and converts each JLD file to an Ovito data file using the write_ovito function.

PeriDyn.jldread — Method

jldread(filename::String)

Reads data from a JLD file.

Arguments:

filename: String, the name of the JLD file to read.

Returns:

Dict, a dictionary containing the data read from the JLD file.

Note: This function uses the load function from the JLD package to read the data from the file, and converts it to a dictionary format.

PeriDyn.load_from_file! — Method

load_from_file!(env::GeneralEnv, filename)

Loads data from a file into the specified environment.

Arguments:

env: GeneralEnv, the environment to load the data into.

filename: String, the name of the file to load data from.

Returns:

Nothing

Note: This function assumes the presence of specific data fields in the file, such as "position", "velocity", and "acceleration".

PeriDyn.make_NN — Method

make_NN(layers::Tuple{T}, N) where T

Create an symmetrical NxN matrix from a vector of length N(N+1)/2.

================

Returns

M :: Matrix

PeriDyn.make_matrix — Method

make_matrix(S::Array{T,1})

Create an symmetrical NxN matrix from a vector of length N(N+1)/2.

================

Returns

M :: Matrix

PeriDyn.make_matrix_gm — Method

make_matrix(S::Array{T,1})

Create an symmetrical NxN matrix from a vector of length N(N+1)/2.

================

Returns

M :: Matrix

PeriDyn.neigh_cells — Method

neigh_cells(i,j,k,N)

Calculate all neighboring cells for i,j,k cell.

PeriDyn.plastic_yield_function — Method

yield function

PeriDyn.print_data_file! — Method

print_data_file!(envs::Array{GeneralEnv}, file_prefix::String, i::Int64)

Writes data file to disk.

PeriDyn.repack! — Method

repack!(d::Dict, keys_, vals)

Repack a dictionary from its components inplace.

Arguments

d::Dict: Dictionary to be repacked.

keys_: Keys of the dictionary.

vals: Values of the dictionary.

Returns

d::Dict: Dictionary containing the components.

PeriDyn.repack — Method

repack(args...; keys_ = (:x, :v, :y, :volume, :type))

Repack a dictionary from its components.

Arguments

args...: Components to be packed.

keys_ = (:x, :v, :y, :volume, :type): Keys of the dictionary.

Returns

d::Dict: Dictionary containing the components.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::LJRepulsionModel11)

Calculates repulsive acceleration for 1-1 materials block interaction.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::LJRepulsionModel12)

Calculates repulsive acceleration for 1-2 materials block interaction.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::NonLinearRepulsionModel11)

Calculates the repulsive acceleration for 1-1 material block interaction.

Arguments:

dr: The vector representing the distance between particles.

RepMod: The NonLinearRepulsionModel11 object.

Returns: The repulsive acceleration as a vector.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::NonLinearRepulsionModel12)

Calculates the repulsive acceleration for 1-2 material block interaction.

Arguments:

dr: The vector representing the distance between particles.

RepMod: The NonLinearRepulsionModel12 object.

Returns: The repulsive acceleration as a vector.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::ShortRangeRepulsionModel11)

Calculates the repulsive acceleration for 1-1 materials block interaction.

Arguments:

dr: Vector{Float64}, the displacement vector between particles.

RepMod: ShortRangeRepulsionModel11, the repulsion model for 1-1 material blocks.

Returns:

Vector{Float64}: The repulsive acceleration vector.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::ShortRangeRepulsionModel12)

Calculates the repulsive acceleration for 1-2 materials block interaction.

Arguments:

dr: Vector{Float64}, the displacement vector between particles.

RepMod: ShortRangeRepulsionModel12, the repulsion model for 1-2 material blocks.

Returns:

Vector{Float64}: The repulsive acceleration vector.

PeriDyn.save_state! — Method

save_state!(filename, env)

Save env::GeneralEnv to disk.

PeriDyn.save_state_ovito_bc! — Method

save_state_ovito_bc!(filename, env)

Save env::GeneralEnv to disk.

PeriDyn.set_env_active! — Method

set_env_active!(env)

Set environment state active.

PeriDyn.set_env_idel! — Method

set_env_idel!(env)

Set environment state idel.

PeriDyn.set_env_inactive! — Method

set_env_inactive!(env)

Set environment state inactive.

PeriDyn.set_ghost_atoms! — Method

set_ghost_atoms!(env,ghost_atoms)

Set ghost atoms for an environment.

PeriDyn.short_range_model_definition — Method

short_range_model_definition(spring_const, del_x, hor, dir)

Calculates the repulsive force using the ShortRange repulsion model formula. The formula is: 18 * springconst / (hor^4 * pi) * (delx / hor) .* dir

Arguments:

spring_const: Float64, the spring constant.

del_x: Float64, the difference in position.

hor: Float64, the characteristic length scale.

dir: Vector{Float64}, the direction vector.

Returns:

Vector{Float64}: The calculated repulsive force vector.

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y, f, type, bid, vol, RM::RepulsionModel11, device::Type{Val{:cpu}})

Updates (inplace) the repulsive acceleration of material points.

Arguments

y: Positions of material points

f: Acceleration of material points

type: Type of material points

bid: BID values

vol: Volume values

RM::RepulsionModel11: Repulsion model

device::Type{Val{:cpu}}: Device type for CPU acceleration

Output

No return value. The function updates f in place.

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y, f, type, bid, vol, RM::RepulsionModel11, device::Type{Val{:cuda}})

Updates (inplace) the repulsive acceleration of material points.

Arguments

y: Positions of material points

f: Acceleration of material points

type: Type of material points

bid: BID values

vol: Volume values

RM::RepulsionModel11: Repulsion model

device::Type{Val{:cuda}}: Device type for CUDA acceleration

Output

No return value. The function updates f in place.

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y, f, type, bid, vol, RM::RepulsionModel11)

Updates (inplace) the repulsive acceleration of material points.

Arguments

y: Positions of material points

f: Acceleration of material points

type: Type of material points

bid: BID values

vol: Volume values

RM::RepulsionModel11: Repulsion model

Output

No return value. The function updates f in place.

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y, f, type, RepulsionModel)

Updates (inplace) the repulsive acceleration of material points.

Arguments

y: Positions of material points.

f: Acceleration of material points.

type: Type of material points.

RepulsionModel: Repulsion model (see contacts.jl for more details).

Output

None (Inplace update of f (acceleration)).

PeriDyn.simulate! — Method

PeriDyn.td_norm_fn — Method

td_norm_fn

PeriDyn.unpack — Method

unpack(d::Dict)

Unpack a dictionary into its components.

Arguments

d::Dict: Dictionary to be unpacked.

Returns

x::Array{Float64, 1}: x coordinates.

v::Array{Float64, 1}: v coordinates.

y::Array{Float64, 1}: y coordinates.

volume::Array{Float64, 1}: Volume of the mesh.

type::Array{Int64, 1}: Type of the mesh.

PeriDyn.update_acc! — Method

update_acc!(env::GeneralEnv)

Updates acceleration of all the material points in a simulation environment.

̈u(xᵢ, t) = from material forces + from contact forces

from material force, ̈u(xᵢ, t) = [∑ᵏⱼ₌₁ {T[xᵢ, t]<xⱼ-xᵢ> - T[xⱼ, t]<xᵢ-xⱼ> }*Vⱼ + b(xᵢ, t)] / ρᵢ

PeriDyn.update_neighs! — Method

function update_neighs!(envs)

Updates neighbors of each material point for a list of simulation environments.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(neighbors, x, search_distance, equi_dist, family, intact, max_part)

Update the neighbor list for repulsive force calculation.

Arguments

neighbors: Array storing the neighbor indices

x: Positions of material points

search_distance: Maximum search distance for neighbors

equi_dist: Equilibrium distance for repulsion

family: Array indicating the family relationship between material points

intact: Array indicating if the family relationship is intact

max_part: Maximum number of particles in a cell

Output

No return value. The function updates the neighbors array in place.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y, type, RM::RepulsionModel11, device::Symbol; kwargs...)

Update neighbor list for repulsive force calculation (1-1 interaction) on a specific device.

Arguments

y: Positions of material points

type: Type of material points

RM::RepulsionModel11: Repulsion model

device::Symbol: Device type for acceleration

kwargs...: Additional keyword arguments

Output

No return value. The function updates RM.neighs in place.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y, type, RM::RepulsionModel11, device::Type{Val{:cpu}}; max_part=30)

Update the neighbor list for repulsive force calculation (1-1 interaction).

Arguments

y: Positions of material points

type: Type of material points

RM::RepulsionModel11: Repulsion model for 1-1 interaction

device::Type{Val{:cpu}}: Device type (CPU)

max_part=30: Maximum number of particles in a cell

Output

No return value. The function updates the neighbor list in the RM object.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y, type, RM::RepulsionModel11, device::Type{Val{:cuda}}; max_part=30)

Update neighbor list for repulsive force calculation (1-1 interaction).

Arguments

y: Positions of material points

type: Type of material points

RM::RepulsionModel11: Repulsion model for 1-1 interaction

device::Type{Val{:cuda}}: Device type (CUDA)

max_part=30: Maximum number of particles in a cell

Output

No return value. The function updates the neighbor list in the RM object.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y, type, RM::RepulsionModel11; kwargs...)

Update neighbor list for repulsive force calculation (1-1 interaction).

Arguments

y: Positions of material points

type: Type of material points

RM::RepulsionModel11: Repulsion model

kwargs...: Additional keyword arguments

Output

No return value. The function updates RM.neighs in place.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y, type, RM::RepulsionModel12; max_part=nothing)

Update neighbor list for repulsive force calculation (1-2 interaction).

Arguments

y: Positions of material points

type: Type of material points

RM::RepulsionModel12: Repulsion model

max_part=nothing: Maximum number of particles (optional)

Output

No return value. The function updates RM.neighs in place.

PeriDyn.velocity_verlet_step! — Method

velocityverletstep!(env::GeneralEnv)

Implement a single step of velocity verlet algorithm.

PeriDyn.weighted_volume — Method

weighted_volume(S::GeneralMaterial)

It gives weighted volume.

PeriDyn.write_data — Method

write_data(filename; kwargs...)

Writes the data file.

Arguments:

filename: String, the name of the file to be written.

kwargs: Keyword arguments, additional options for writing the file.

Note: This function supports writing files in the .data and .jld formats.

PeriDyn.write_global_data — Method

write_global_data(filename; kwargs...)

Writes the global data file.

Arguments:

filename: String, the name of the file to be written.

kwargs: Keyword arguments, additional options for writing the file.

Note: This function writes a data file in the .jld format.

PeriDyn.write_ovito — Method

write_ovito(filename::String; kwargs...)

Writes data to an Ovito data file.

Arguments:

filename: String, the name of the file to write.

kwargs: Keyword arguments, the data to write to the file.

Note: This function writes data in Ovito data file format, with each column specified by a keyword argument.

PeriDyn.write_ovito_cell_ids — Method

write_ovito_cell_ids(filename::String, y::Matrix, horizon)

Writes cell IDs to an Ovito data file.

Arguments:

filename: String, the name of the file to write.

y: Matrix, the coordinates of the particles.

horizon: The horizon value.

Note: This function writes the cell IDs to an Ovito data file, based on the particle coordinates and horizon value.