PeriDyn.jl Documentation

Simulation Environment

PeriDyn.Env — Method

Env(id::Int64,materials,short_range_repulsion,boundary_conds,dt;state=2)

Create a GeneralEnv for holding parameters for a simulation.

Mesh

PeriDyn.create_block — Method

create_block(lattice::Array{Float64,1}, N::Array{Int64,1})::Array{Float64,2}

Create a rectangular point mesh.

Boundary Conditions

PeriDyn.apply_bc! — Method

apply_bc!(env,BC::FixBC)

Applies fixed boundary condition to a given material block.

PeriDyn.apply_bc! — Method

apply_bc!(env,BC::MoveBC)

Applies moving boundary condition to a given material block. Material will move with the given rate (constant velocity).

PeriDyn.apply_bc! — Method

apply_bc!(env,BC::ScaleBC)

It scales a given material block about a given fixed point.

PeriDyn.apply_bc! — Method

apply_bc!(env,BC::JustScaleBC)

It scales a given material block about a given fixed point.

PeriDyn.apply_bc! — Method

apply_bc!(env,BC::ScaleFixBC)

It scales as well as fix a given material block. (like scale and fix edges, helpful in tensile quasi-static simulations.)

Solvers

PeriDyn.velocity_verlet! — Method

velocity_verlet!(envs::Any,N::Int64;filewrite_freq=10,neigh_update_freq=50,file_prefix="datafile",start_at::Int64=0)

Velocity verlet :).

PeriDyn.minimize! — Method

minimize!(env::GeneralEnv,step_size::Float64; max_iter::Int64=500,min_step_tol_per::Float64=5.0)

Minimize potential energy of simulation environment.

PeriDyn.quasi_static! — Method

quasi_static!(envs::Any,N::Int64,step_size::Float64; max_iter::Int64=100, min_step_tol_per::Float64=0.5, filewrite_freq::Int64=10, neigh_update_freq::Int64=50, file_prefix::String="datafile",start_at::Int64=0)

Implements quasi static simulation using minimize for each step.

Util

PeriDyn.write_data — Method

write_data(filename,every,type,pos)

It writes the data file.

PeriDyn.write_data — Method

write_data(filename::String,type::Array{Int64,1},y::Array{Float64,2})

It writes the data file.

PeriDyn.write_data — Method

write_data(filename::String,type::Array{Int64,1},y::Array{Float64,2}, v::Array{Float64,2},f::Array{Float64,2},)

It writes the data file.