PeriDyn.BondBasedMaterial — Type

Bond based material type.

PeriDyn.BondBasedSpecific — Type

Specific bond based material type.

PeriDyn.BoundaryCondition — Type

Abstract class for boundary conditions.

PeriDyn.GeneralMaterial — Method

GeneralMaterial(y0,v0,x,volume,density,horizon,critical_stretch; particle_size=0,max_neigh=50)

General peridynamics material type.

PeriDyn.OrdinaryStateBasedSpecific — Type

Specific ordinary state based materail type.

PeriDyn.OrdinaryStateBasedSpecific — Method

OrdinaryStateBasedSpecific(bulk_modulus::Float64, shear_modulus::Float64, mat_gen::GeneralMaterial)

Specific ordinary state based materail type.

PeriDyn.PeridynamicsMaterial — Type

Abstract PeridynamicsMaterial type.

PeriDyn.SimpleRepulsionModel — Method

SimpleRepulsionModel(alpha::Float64,epsilon::Float64, mat1::PeridynamicsMaterial,mat2::PeridynamicsMaterial;distanceX=5,max_neighs=50)

Simple repulsive model for 1-2 material blocks.

PeriDyn.SimpleRepulsionModel — Method

SimpleRepulsionModel(alpha::Float64,epsilon::Float64, mat1::PeridynamicsMaterial;distanceX=5,max_neighs=50)

Simple repulsive model for 1-1 material blocks.

PeriDyn.Xij — Method

X(x::Array{Float64,1})

Peridynamics state: Vector {i,j} operator

PeriDyn._tij — Method

_tij(x,mi,thetai,eij,K,G)::Float64

calculates tij force density magnitude (actually acceleration).

PeriDyn.cal_family! — Method

cal_family!(family::Array{Int64,2},x::Array{Float64,2}, horizon::Float64)

Calculate family members for each material point (inplace).

PeriDyn.cal_family — Method

cal_family!(family::Array{Int64,2},x::Array{Float64,2}, horizon::Float64)

Calculate family members for each material point.

PeriDyn.cell_number — Method

cell_number(i,j,k,N)

Calculates cell number for given i,j,k cell indices (when in a list).

PeriDyn.collision_box — Method

collision_box(x1::Array{Float64,2}, x2::Array{Float64,2}, skin::Float64)

Calculates collsion box between two material blocks.

Input Args:

x1 :: Positions of material point (block 1)

x2 :: Positions of material point (block 2)

skin :: Extra distance need to consider (usually >= particle size)

Output Args:

box_min :: Minimum position limits for overlap

box_max :: Miximum position limits for overlap

ifoverlap :: Boolean (true if overlap)

Examples

julia> collision_box(x1, x2, skin)

PeriDyn.dilatation — Method

dilatation(y::Array{Float64,2},S::GeneralMaterial,m::Array{Float64,1})

It gives dilatation as given ordinary state material model.

PeriDyn.dilatation_theta — Method

dilatation_theta(y::Array{Float64,2},S::GeneralMaterial)

It gives dilatation(theta) as given ordinary state material model.

PeriDyn.force_density_T — Method

force_density_T(mat::BondBasedMaterial)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(y::Array{Float64,2},mat::OrdinaryStateBasedMaterial)

Calculates force density (actually acceleration) for ordinary state based material type.

PeriDyn.get_cells — Method

get_cells(x::Array{Float64,2},horizon::Float64)

Fill cells with material points.

PeriDyn.horizon_correction — Method

horizon_correction(dr,r,hr)

It gives horizon correction factor (It will give 1 as of now).

PeriDyn.identity — Method

identity(x::Array{Float64,1})

Peridynamics state: Identity operator

PeriDyn.influence_function — Method

influence_function(dr)

It gives influence function factor (It will give 1/r as of now).

PeriDyn.magnitude — Method

magnitude(x::Array{Float64,1})

Peridynamics state: Modulus operator

PeriDyn.neigh_cells — Method

neigh_cells(i,j,k,N)

Calculate all neighboring cells for i,j,k cell.

PeriDyn.null — Method

null(x::Array{Float64,1})

Peridynamics state: Null operator

PeriDyn.print_data_file! — Method

print_data_file!(envs::Array{GeneralEnv}, file_prefix::String,i::Int64)

Writes data file to disk.

PeriDyn.repulsion_acc — Method

repulsion_acc(dr,den_i,RepMod::SimpleRepulsionModel12)

Calculates repulsive acceleration for 1-2 materials block interaction.

PeriDyn.repulsion_acc — Method

repulsion_acc(dr,den_i,RepMod::SimpleRepulsionModel11)

Calculates repulsive acceleration for 1-1 materials block interaction.

PeriDyn.set_env_active! — Method

set_env_active!(env)

Set environment state active.

PeriDyn.set_env_idel! — Method

set_env_idel!(env)

Set environment state idel.

PeriDyn.set_env_inactive! — Method

set_env_inactive!(env)

Set environment state inactive.

PeriDyn.set_ghost_atoms! — Method

set_ghost_atoms!(env,ghost_atoms)

Set ghost atoms for an environment.

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y,f,type,RepusionModel)

Updates (inplace) the repulsive acceleration of material points.

1-2 repulsion

Input Args:

y :: Positions of material point

f :: Acceleration of material points

type :: Type of material points

RepulsionModel :: Repulsion model (see contacts.jl for more details)

Output Args:

Noting (Inplace updation of f (acceleration))

Examples

julia> short_range_repulsion!(y,f,type,RepusionModel)

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y,f,type,RepusionModel)

Updates (inplace) the repulsive acceleration of material points.

$α$

1-1 repulsion

Input Args:

y :: Positions of material point

f :: Acceleration of material points

type :: Type of material points

RepulsionModel :: Repulsion model (see contacts.jl for more details)

Output Args:

Noting (Inplace updation of f (acceleration))

Examples

julia> short_range_repulsion!(y,f,type,RepusionModel)

PeriDyn.update_acc! — Method

update_acc!(env::GeneralEnv)

Updates acceleration of all the material points in a simulation environment.

PeriDyn.update_neighs! — Method

function update_neighs!(envs)

Updates neighbors of each material point for a list of simulation environments.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y,type,RepulsionModel11)

Update neighbour list for repulsive force calculation (1-1 interaction).

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y,type,RepulsionModel12)

Update neighbour list for repulsive force calculation (1-2 interaction).

PeriDyn.velocity_verlet_step! — Method

velocity_verlet_step!(env::GeneralEnv)

Implement a single step of velocity verlet algorithm.

PeriDyn.weighted_volume — Method

weighted_volume(S::GeneralMaterial)

It gives weighted volume.

PeriDyn.write_data_cell_ids — Method

write_data_cell_ids(filename::String,y::Array{Float64,2},cells::Any)

It writes the data file.