Solvers

PeriDyn.Solver — Type

Solver

Solver is an abstract type for solvers.

PeriDyn.QuasiStaticSolver — Type

QuasiStaticSolver <: Solver

QuasiStaticSolver is an abstract type for quasi-static solvers.

PeriDyn.DynamicSolver — Type

DynamicSolver <: Solver

DynamicSolver is an abstract type for dynamic solvers.

PeriDyn.apply_solver! — Function

apply_solver!(env, solver::DSVelocityVerlet)

Apply the velocity verlet algorithm to the environment.

Arguments

env: GeneralEnv, the environment.

solver: DSVelocityVerlet, the velocity verlet algorithm.

Example

solver = DSVelocityVerlet()
apply_solver!(env, solver)

apply_solver!(env, solver::QSDrag)

Apply the QSDrag solver to the environment.

Arguments

env: GeneralEnv, the environment.

solver: QSDrag, the QSDrag solver.

Example

solver = QSDrag(1.0e-3, 1.0e-3)
apply_solver!(env, solver)

QuasiStaticSolver

PeriDyn.QSDrag — Type

QSDrag

Quasi-static solver struct.

Fields

step_size: Float64, the step size.

drag: Float64, the drag coefficient.

max_iter: Int64, the maximum number of iterations. Default is 1000.

x_tol: Float64, the tolerance for position. Default is 1.0e-3.

f_tol: Float64, the tolerance for force. Default is 1.0e-3.

Example

solver = QSDrag(1.0e-3, 1.0e-3)

See also

QSDrag

apply_solver!

minimize!

PeriDyn.apply_solver! — Method

apply_solver!(env, solver::QSDrag)

Apply the QSDrag solver to the environment.

Arguments

env: GeneralEnv, the environment.

solver: QSDrag, the QSDrag solver.

Example

solver = QSDrag(1.0e-3, 1.0e-3)
apply_solver!(env, solver)

PeriDyn.minimize! — Function

minimize!(env, solver::QSDrag)

Minimize the environment using the QSDrag solver.

Arguments

env: GeneralEnv, the environment.

solver: QSDrag, the QSDrag solver.

Example

solver = QSDrag(1.0e-3, 1.0e-3)
minimize!(env, solver)

DynamicSolver

PeriDyn.DSVelocityVerlet — Type

DSVelocityVerlet

The velocity verlet algorithm.

PeriDyn.apply_solver! — Method

apply_solver!(env, solver::DSVelocityVerlet)

Apply the velocity verlet algorithm to the environment.

Arguments

env: GeneralEnv, the environment.

solver: DSVelocityVerlet, the velocity verlet algorithm.

Example

solver = DSVelocityVerlet()
apply_solver!(env, solver)

PeriDyn.velocity_verlet_step! — Function

velocity_verlet_step!(env, solver::DSVelocityVerlet)

Apply one step of the velocity verlet algorithm to the environment.

Arguments

env: GeneralEnv, the environment.

solver: DSVelocityVerlet, the velocity verlet algorithm.

Example

solver = DSVelocityVerlet()
velocity_verlet_step!(env, solver)

Running the simulation

PeriDyn.simulate! — Function

simulate!(args...; out_dir="datafile", append_date=true, kwargs...)

Simulate a list of environments. The last argument should be a solver. The args and kwargs are passed to run! function. The out_dir is the directory where the data files are saved. The append_date is a boolean value. If true, the date is appended to the out_dir path.

PeriDyn.run! — Function

run!(envs, N::Int64, solver::Solver; filewrite_freq::Int64=10,
    neigh_update_freq::Int64=1, average_prop_freq::Int64=1,
    out_dir::String="datafile",
    start_at::Int64=0, write_from::Int=0, ext::Symbol=:jld,
    max_part=30)

Run simulation for a list of environments.

Arguments

envs: Array{GeneralEnv}, the list of environments.

N: Int64, the number of steps.

solver: Solver, the solver.

Keyword Arguments

filewrite_freq: Int64, the frequency of writing data files to disk. Default is 10.

neigh_update_freq: Int64, the frequency of updating neighbors. Default is 1.

average_prop_freq: Int64, the frequency of calculating average properties. Default is 1.

out_dir: String, the directory where the data files are saved. Default is "datafile".

start_at: Int64, the starting step. Default is 0.

write_from: Int, the starting index of the data files. Default is 0.

ext: Symbol, the extension of the data files. Default is :jld.

max_part: Int, the maximum number of particles in a neighborhood. Default is 30.

See also

simulate!

PeriDyn.update_acc! — Function

update_acc!(env::GeneralEnv)

Updates acceleration of all the material points in a simulation environment.

̈u(xᵢ, t) = forces due to material deformation + forces due to contact

forces due to material deformation: > ̈u(xᵢ, t) = [∑ᵏⱼ₌₁ {T[xᵢ, t]<xⱼ-xᵢ> - T[xⱼ, t]<xᵢ-xⱼ> }*Vⱼ + b(xᵢ, t)] / ρᵢ

Arguments

env: GeneralEnv, the simulation environment.

See also

force_density

short_range_repulsion!

PeriDyn.update_neighs! — Function

function update_neighs!(envs)

Updates neighbors of each material point for a list of simulation environments.

Arguments

envs: Array{GeneralEnv}, the list of simulation environments.

See also

update_repulsive_neighs!

PeriDyn.apply_bc_at0 — Function

apply_bc_at0!(env, start_at)

Apply boundary conditions at t = t0.

Arguments

env: GeneralEnv, the simulation environment.

start_at: Int64, the starting step.

Saving the simulation state

PeriDyn.filepath_ — Function

filepath_(file_prefix::String; append_date=false)

Returns the path to the output folder. If append_date is true, the date is appended.

Arguments

file_prefix: String, the prefix of the output folder.

Keyword Arguments

append_date: Bool, the boolean value to append date to the output folder. Default is

false.

PeriDyn.save_state! — Function

save_state!(filename, env)

Save env::GeneralEnv to disk.

Arguments

filename: String, the filename.

env: GeneralEnv, the simulation environment.

See also

save_state_ovito_bc!

PeriDyn.save_state_ovito_bc! — Function

save_state_ovito_bc!(filename, env)

Save env::GeneralEnv to disk for ovito visualization. The boundary conditions are saved as type.

Arguments

filename: String, the filename.

env: GeneralEnv, the simulation environment.

See also

save_state!

PeriDyn.print_data_file! — Function

print_data_file!(envs::Array{GeneralEnv}, file_prefix::String, i::Int64)

Prints data files to disk for a list of simulation environments.

Arguments

envs: Array{GeneralEnv}, the list of simulation environments.

file_prefix: String, the prefix of the output folder.

i: Int64, the step number.