PeriDyn.BondBasedMaterial — Type

Bond based material type.

PeriDyn.BondBasedSpecific — Type

Specific bond based material type.

PeriDyn.BoundaryCondition — Type

Abstract class for boundary conditions.

PeriDyn.ElastoPlasticSolidMaterial — Type

ElastoPlasticSolidMaterial(type::UnitRange{Int64}, general::GeneralMaterial, specific::ElastoPlasticSolidSpecific)

PeriDyn.ElastoPlasticSolidSpecific — Type

Specific Elasto Plastic Solid Material type.

PeriDyn.ElastoPlasticSolidSpecific — Method

ElastoPlasticSolidSpecific(bulk_modulus::Array{Float64, 1}, shear_modulus::Array{Float64,1}, critical_stretch::Array{Float64,1}, density::Array{Float64,1})

Specific Elasto Plastic Solid Material type.

PeriDyn.ElastoPlasticSolidSpecificFull — Type

Specific Elasto Plastic Solid Material type.

PeriDyn.FixBC — Method

FixBC(bool; onlyatstart=false)

PeriDyn.GeneralMaterial — Method

GeneralMaterial(y0,v0,x,volume,type,horizon; particle_size=0,max_neigh=50)

General peridynamics material type.

PeriDyn.GeneralMaterial — Method

PeriDyn.OrdinaryStateBasedMaterial — Type

OrdinaryStateBasedMaterial(type::UnitRange{Int64}, general::GeneralMaterial, specific::OrdinaryStateBasedSpecific)

PeriDyn.OrdinaryStateBasedSpecific — Type

Specific ordinary state based materail type.

PeriDyn.OrdinaryStateBasedSpecific — Method

OrdinaryStateBasedSpecific(bulk_modulus::Array{Float64, 1}, shear_modulus::Array{Float64,1}, critical_stretch::Array{Float64,1}, density::Array{Float64,1})

Specific ordinary state based materail type.

PeriDyn.PairwiseNNMaterial — Type

Bond based material type (NN).

PeriDyn.PairwiseNNSpecific — Type

Specific bond based material type (NN) .

PeriDyn.PeridynamicsMaterial — Type

Abstract PeridynamicsMaterial type.

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(gen, spc)

Create peridynamics material.

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(gen, spc::ElastoPlasticSolidSpecific)

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(gen, spc::SkipSpecific)

PeriDyn.PeridynamicsMaterial — Method

PeridynamicsMaterial(gen, spc::PeridynamicsMaterial; name="PM")

PeriDyn.SkipMaterial — Type

SkipMaterial(type::UnitRange{Int64}, general::GeneralMaterial, specific::SkipSpecific)

PeriDyn.SkipSpecific — Type

Specific skip materail type.

PeriDyn.SkipSpecific — Method

Specific skip materail type.

PeriDyn.Env — Method

Env(id::Int64,materials,short_range_repulsion,boundary_conds,dt;state=2)

Create a GeneralEnv for holding parameters for a simulation.

PeriDyn.LJRepulsionModel — Method

LJRepulsionModel(alpha::Float64,epsilon::Float64, mat1::PeridynamicsMaterial,mat2::PeridynamicsMaterial;distanceX=5,max_neighs=50)

LJ repulsive model for 1-2 material blocks.

PeriDyn.LJRepulsionModel — Method

LJRepulsionModel(alpha::Float64,epsilon::Float64, mat1::PeridynamicsMaterial;distanceX=5,max_neighs=50)

LJ repulsive model for 1-1 material blocks.

PeriDyn.LinearRepulsionModel — Method

LinearRepulsionModel(stifness::Float64, mat1::PeridynamicsMaterial,mat2::PeridynamicsMaterial;distanceX=5,max_neighs=50)

Linear repulsive model for 1-2 material blocks.

PeriDyn.LinearRepulsionModel — Method

LinearRepulsionModel(stifness::Float64, mat1::PeridynamicsMaterial;distanceX=5,max_neighs=50)

Linear repulsive model for 1-1 material blocks.

PeriDyn.NonLinearRepulsionModel — Method

NonLinearRepulsionModel(exponent::Float64,stifness::Float64, mat1::PeridynamicsMaterial,mat2::PeridynamicsMaterial;distanceX=5,max_neighs=50)

NonLinear repulsive model for 1-2 material blocks.

PeriDyn.NonLinearRepulsionModel — Method

NonLinearRepulsionModel(exponent::Float64,stifness::Float64, mat1::PeridynamicsMaterial;distanceX=5,max_neighs=50)

NonLinear repulsive model for 1-1 material blocks.

PeriDyn.ShortRangeRepulsionModel — Method

ShortRangeRepulsionModel(spring_const::Float64, mat1::PeridynamicsMaterial,mat2::PeridynamicsMaterial;distanceX=5,max_neighs=50)

ShortRange repulsive model for 1-2 material blocks.

PeriDyn.ShortRangeRepulsionModel — Method

ShortRangeRepulsionModel(spring_const::Float64, mat1::PeridynamicsMaterial;distanceX=5,max_neighs=50)

ShortRange repulsive model for 1-1 material blocks.

PeriDyn._tij — Method

_tij(x,mi,thetai,eij,K,G)::Float64

calculates tij force density magnitude (actually acceleration).

PeriDyn.apply_bc! — Method

apply_bc!(env,BC::BoundaryCondition, :position)

Applies general boundary condition to a given material block.

PeriDyn.apply_bc! — Method

apply_bc!(env,BC::BoundaryCondition, :velocity)

Applies general boundary condition to a given material block.

PeriDyn.cal_family! — Method

cal_family!(family::Array{Int64,2},x::Array{Float64,2}, horizon::Float64)

Calculate family members for each material point (inplace).

PeriDyn.cal_family — Method

cal_family!(family::Array{Int64,2},x::Array{Float64,2}, horizon::Float64)

Calculate family members for each material point.

PeriDyn.cell_number — Method

cell_number(i,j,k,N)

Calculates cell number for given i,j,k cell indices (when in a list).

PeriDyn.collision_box — Method

collision_box(x1::Array{Float64,2}, x2::Array{Float64,2}, skin::Float64)

Calculates collsion box between two material blocks.

Input Args:

x1 :: Positions of material point (block 1)

x2 :: Positions of material point (block 2)

skin :: Extra distance need to consider (usually >= particle size)

Output Args:

box_min :: Minimum position limits for overlap

box_max :: Miximum position limits for overlap

ifoverlap :: Boolean (true if overlap)

Examples

julia> collision_box(x1, x2, skin)

PeriDyn.dilatation — Method

dilatation(y::Array{Float64,2},S::GeneralMaterial,m::Array{Float64,1})

It gives dilatation as given ordinary state material model.

PeriDyn.dilatation_theta — Method

dilatation_theta(y::Array{Float64,2},S::GeneralMaterial)

It gives dilatation as given ordinary state material model.

PeriDyn.force_density_T — Method

force_density_T(mat::BondBasedMaterial)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(y::Array{Float64,2},mat::ElastoPlasticSolidMaterial)

Calculates force density (actually acceleration) for ordinary state based material type.

PeriDyn.force_density_T — Method

force_density_T(y::Array{Float64,2},mat::OrdinaryStateBasedMaterial)

Calculates force density (actually acceleration) for ordinary state based material type.

PeriDyn.force_density_T — Method

force_density_T(mat::PairwiseNNMaterial)

Calculates force density (actually acceleration) for bond based material type.

PeriDyn.force_density_T — Method

force_density_T(y::Array{Float64,2},mat::SkipMaterial)

Calculates force density (actually acceleration) for skip material type.

PeriDyn.get_cells — Method

get_cells(x::Array{Float64,2},horizon::Float64)

Fill cells with material points.

PeriDyn.horizon_correction — Method

horizon_correction(dr,r,hr)

It gives horizon correction factor (It will give 1 as of now).

PeriDyn.influence_function — Method

influence_function(dr)

It gives influence function factor (It will give 1/r as of now).

PeriDyn.make_NN — Method

make_NN(layers::Tuple{T}, N) where T

Create an symmetrical NxN matrix from a vector of length N(N+1)/2.

================

Returns

M :: Matrix

PeriDyn.make_matrix — Method

make_matrix(S::Array{T,1})

Create an symmetrical NxN matrix from a vector of length N(N+1)/2.

================

Returns

M :: Matrix

PeriDyn.make_matrix_gm — Method

make_matrix(S::Array{T,1})

Create an symmetrical NxN matrix from a vector of length N(N+1)/2.

================

Returns

M :: Matrix

PeriDyn.neigh_cells — Method

neigh_cells(i,j,k,N)

Calculate all neighboring cells for i,j,k cell.

PeriDyn.plastic_yield_function — Method

yield function

PeriDyn.print_data_file! — Method

print_data_file!(envs::Array{GeneralEnv}, file_prefix::String, i::Int64)

Writes data file to disk.

PeriDyn.quasi_static! — Method

quasi_static!(envs::Any,N::Int64,step_size::Float64; max_iter::Int64=100, min_step_tol_per::Float64=0.5, filewrite_freq::Int64=10, neigh_update_freq::Int64=50, file_prefix::String="datafile",start_at::Int64=0,write_from::Int=0)

Implements quasi static simulation using minimize for each step.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::LJRepulsionModel11)

Calculates repulsive acceleration for 1-1 materials block interaction.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::LJRepulsionModel12)

Calculates repulsive acceleration for 1-2 materials block interaction.

PeriDyn.repulsion_force — Method

repulsion_force(dr,RepMod::NonLinearRepulsionModel11)

Calculates repulsive acceleration for 1-1 materials block interaction.

PeriDyn.repulsion_force — Method

repulsion_force(dr,RepMod::NonLinearRepulsionModel12)

Calculates repulsive acceleration for 1-2 materials block interaction.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::ShortRangeRepulsionModel11)

Calculates repulsive acceleration for 1-1 materials block interaction.

PeriDyn.repulsion_force — Method

repulsion_force(dr, RepMod::ShortRangeRepulsionModel12)

Calculates repulsive acceleration for 1-2 materials block interaction.

PeriDyn.save_state! — Method

save_state!(filename, env)

Save env::GeneralEnv to disk.

PeriDyn.set_env_active! — Method

set_env_active!(env)

Set environment state active.

PeriDyn.set_env_idel! — Method

set_env_idel!(env)

Set environment state idel.

PeriDyn.set_env_inactive! — Method

set_env_inactive!(env)

Set environment state inactive.

PeriDyn.set_ghost_atoms! — Method

set_ghost_atoms!(env,ghost_atoms)

Set ghost atoms for an environment.

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y,f,type,RepusionModel)

Updates (inplace) the repulsive acceleration of material points.

PeriDyn.short_range_repulsion! — Method

short_range_repulsion!(y,f,type,RepusionModel)

Updates (inplace) the repulsive acceleration of material points.

$α$

1-1 repulsion

Input Args:

y :: Positions of material point

f :: Acceleration of material points

type :: Type of material points

RepulsionModel :: Repulsion model (see contacts.jl for more details)

Output Args:

Noting (Inplace updation of f (acceleration))

Examples

julia> short_range_repulsion!(y,f,type,RepusionModel)

PeriDyn.simulate — Method

PeriDyn.td_norm_fn — Method

td_norm_fn

PeriDyn.update_acc! — Method

update_acc!(env::GeneralEnv)

Updates acceleration of all the material points in a simulation environment.

̈u(xᵢ, t) = from material forces + from contact forces

from material force, ̈u(xᵢ, t) = [∑ᵏⱼ₌₁ {T[xᵢ, t]<xⱼ-xᵢ> - T[xⱼ, t]<xᵢ-xⱼ> }*Vⱼ + b(xᵢ, t)] / ρᵢ

PeriDyn.update_neighs! — Method

function update_neighs!(envs)

Updates neighbors of each material point for a list of simulation environments.

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y,type,RepulsionModel11)

Update neighbour list for repulsive force calculation (1-1 interaction).

PeriDyn.update_repulsive_neighs! — Method

update_repulsive_neighs!(y,type,RepulsionModel12)

Update neighbour list for repulsive force calculation (1-2 interaction).

PeriDyn.velocity_verlet! — Method

velocityverlet!(envs::Any,N::Int64;filewritefreq=10,neighupdatefreq=50,fileprefix="datafile",startat::Int64=0,write_from::Int=0)

Velocity verlet :).

PeriDyn.velocity_verlet_step! — Method

velocityverletstep!(env::GeneralEnv)

Implement a single step of velocity verlet algorithm.

PeriDyn.weighted_volume — Method

weighted_volume(S::GeneralMaterial)

It gives weighted volume.

PeriDyn.write_data — Method

write_data(filename,every,type,pos)

It writes the data file.

PeriDyn.write_data — Method

write_data(filename::String, args...)

It writes the data file.

PeriDyn.write_data — Method

write_data(filename::String,type::Array{Int64,1},y::Array{Float64,2}, v::Array{Float64,2},f::Array{Float64,2},)

It writes the data file.

PeriDyn.write_data — Method

write_data(filename::String,type::Array{Int64,1},y::Array{Float64,2})

It writes the data file.

PeriDyn.write_data_cell_ids — Method

write_data_cell_ids(filename::String, y::Array{Float64,2}, cells::Any)

It writes the data file.

PDMesh.Cone — Type

Cone

Shape object of cone.

Args

- radius : radius of cone
- length : length of cone

Returns

- obj : Shape object **Cone**.

PDMesh.PostOpObj — Type

PostOpObj

Create a post operation object (lazy) from object ( or post operation object) and an operation. Operations will be applied while a create call.

Returns

- obj : Post operation object (lazy)

PDMesh.Shape — Type

Shape
Abstarct shape object.

PDMesh.create — Method

create(shape::T; resolution=nothing, rand_=0.0, type::Int64=1) where T <: Shape

Abstact function for creating Shape objects.

Returns

- X : Initial reference position 
- V : Initial velocity 
- Y : Initial position 
- volume : Volume per particle point 
- type : Type of particle point

PDMesh.delete — Method

delete(obj::T, f::Function) where T

Delete mesh particle for object using boolean array from function f.