Boundary conditions

PeriDyn.BoundaryCondition — Type

BoundaryCondition

Abstract type for boundary conditions.

PeriDyn.apply_bc! — Function

apply_bc!(env, BC::BoundaryCondition, on::Symbol)

Apply the specified boundary condition BC to the given environment env on the specified aspect on.

Arguments

env: The environment to which the boundary condition is applied.

BC: The boundary condition to apply.

on::Symbol: The aspect on which the boundary condition is applied (:position or :velocity).

apply_bc!(env, BC::BoundaryCondition, ::Type{Val{:position}})

Apply the general boundary condition BC to the position aspect of the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The general boundary condition to apply.

apply_bc!(env, BC::BoundaryCondition, ::Type{Val{:velocity}})

Apply the general boundary condition BC to the velocity aspect of the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The general boundary condition to apply.

PeriDyn.apply_bc_at0! — Function

apply_bc_at0!(env, BC::BoundaryCondition)

Apply the boundary condition BC to the given environment env at the start of the simulation (t=0).

apply_bc_at0!(env, BC::FixBC)

Apply the FixBC boundary condition at time 0 to the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The FixBC boundary condition to apply.

apply_bc_at0!(env, BC::ToFroBC)

Apply the ToFroBC boundary condition at time 0 to the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The ToFroBC boundary condition to apply.

apply_bc_at0!(env, BC::DeltaScaleBC)

Apply the DeltaScaleBC boundary condition at time 0 to the given environment env.

Arguments

env: The environment to which the boundary condition is applied.

BC: The DeltaScaleBC boundary condition to apply.

apply_bc_at0!(env, BC::ScaleFixWaitBC)

Apply the ScaleFixWaitBC boundary condition at time 0.

Arguments

env: Environment in which the condition is applied.

BC: Instance of ScaleFixWaitBC boundary condition.

PeriDyn.check! — Function

check!(env, BC::BoundaryCondition)

Check if the boundary condition BC has changed and updates the BC. Used for dynamic boundary conditions.

check!(BC::ToFroBC, env)

Perform a check on the ToFroBC boundary condition.

Arguments

BC: The ToFroBC boundary condition to check.

env: The environment associated with the boundary condition.

check!(BC::DeltaScaleBC, env)

Perform a check on the DeltaScaleBC boundary condition.

Arguments

BC: The DeltaScaleBC boundary condition to check.

env: The environment associated with the boundary condition.

check!(BC::ScaleFixWaitBC, env)

Check if the ScaleFixWaitBC boundary condition needs to be applied.

Arguments

BC: Instance of ScaleFixWaitBC boundary condition.

env: Environment in which the condition is applied.

PeriDyn.FixBC — Type

FixBC

Struct representing the FixBC boundary condition.

Fields

bool: Boolean array specifying the affected elements.

last: Last position of the affected elements.

onlyatstart: Flag indicating if the boundary condition is applied only at the start.

xF: Function for updating the velocity.

vF: Function for updating the position.

PeriDyn.FixBC — Method

FixBC(bool; onlyatstart=false)

Construct a FixBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

onlyatstart: Flag indicating if the boundary condition is applied only at the start (default: false).

Returns

A FixBC object representing the boundary condition.

PeriDyn.ToFroBC — Type

ToFroBC

Struct representing the ToFroBC boundary condition.

Fields

bool: Boolean array specifying the affected elements.

last: Last position of the affected elements.

onlyatstart: Flag indicating if the boundary condition is applied only at the start.

xF: Function for updating the velocity.

vF: Function for updating the position.

direction: Direction of movement.

freq: Frequency at which the direction of movement changes.

applyafter: Number of steps after which the frequency is applied.

PeriDyn.ToFroBC — Method

ToFroBC(bool, rate, freq; applyafter=0, onlyatstart=false)

Construct a ToFroBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

rate: Rate at which the elements move.

freq: Frequency at which the direction of movement changes.

applyafter: Number of steps after which the frequency is applied (default: 0).

onlyatstart: Flag indicating if the boundary condition is applied only at the start (default: false).

Returns

A ToFroBC object representing the boundary condition.

PeriDyn.MoveBC — Function

MoveBC(bool, rate; kwargs...)

Create a MoveBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

rate: Rate at which the elements move.

kwargs: Additional keyword arguments passed to ToFroBC.

Returns

A MoveBC object representing the boundary condition. All MoveBC objects are ToFroBC objects with frequency Inf.

PeriDyn.DeltaScaleBC — Type

DeltaScaleBC

Struct representing the DeltaScaleBC boundary condition.

Fields

bool: Boolean array specifying the affected elements.

last: Last position of the affected elements.

onlyatstart: Flag indicating if the boundary condition is applied only at the start.

xF: Function for updating the velocity.

vF: Function for updating the position.

PeriDyn.DeltaScaleBC — Method

DeltaScaleBC(bool, scale, fixpoint; onlyatstart=false)

Construct a DeltaScaleBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

scale: Scale factor applied to the elements.

fixpoint: Reference point used for scaling.

onlyatstart: Flag indicating if the boundary condition is applied only at the start (default: false).

Returns

A DeltaScaleBC object representing the boundary condition.

PeriDyn.ScaleFixWaitBC — Type

ScaleFixWaitBC

Structure representing a ScaleFixWaitBC boundary condition.

Fields

bool: Boolean array specifying the affected elements.

last: Last position of the affected elements.

onlyatstart: Flag indicating if the boundary condition is applied only at the start.

xF: Function for updating the velocity.

vF: Function for updating the position.

checkF: Function for checking if the boundary condition needs to be applied.

PeriDyn.ScaleFixWaitBC — Method

ScaleFixWaitBC(bool, scale, fixpoint, wait, scalebool; applyafter=0, onlyatstart=false)

Construct a ScaleFixWaitBC boundary condition.

Arguments

bool: Boolean array specifying the affected elements.

scale: Scale factor for the elements.

fixpoint: Fix point for the elements.

wait: Number of time steps to wait before applying the condition.

scalebool: Boolean array specifying the elements to be scaled.

applyafter: Number of time steps after which the condition is applied (default: 0).

onlyatstart: Boolean indicating whether the condition is applied only at the start (default: false).

Returns

An instance of ScaleFixWaitBC boundary condition.